refactor: приведение к единому стилю кода
