+++ src/global/render.js
@@ -5,15 +5,15 @@ const RenderPosition = {
   AFTEREND: 'afterend',
 };
 
-function createElement(template) {
+const createElement = (template) => {
   const newElement = document.createElement('div');
   newElement.innerHTML = template;
 
   return newElement.firstElementChild;
-}
+};
 
-function render(component, container, place = RenderPosition.BEFOREEND) {
+const render = (component, container, place = RenderPosition.BEFOREEND) => {
   container.insertAdjacentElement(place, component.getElement());
-}
+};
 
 export {RenderPosition, createElement, render};
