+++ src/modules/events/view/events-sort-view.js
@@ -1,7 +1,6 @@
-import { createElement } from '../../render.js';
+import { createElement } from '../../../global/render.js';
 
-function createEventsSortTemplate() {
-  return (/*html*/`
+const createEventsSortTemplate = () => (/*html*/`
     <form class="trip-events__trip-sort  trip-sort" action="#" method="get">
       <div class="trip-sort__item  trip-sort__item--day">
         <input id="sort-day" class="trip-sort__input  visually-hidden" type="radio" name="trip-sort" value="sort-day" checked>
@@ -28,8 +27,7 @@ function createEventsSortTemplate() {
         <label class="trip-sort__btn" for="sort-offer">Offers</label>
       </div>
     </form>`
-  );
-}
+);
 
 export default class EventsSortView {
   getTemplate() {
