+++ src/modules/trip-main/view/trip-main-info-view.js
@@ -1,7 +1,6 @@
-import { createElement } from '../../render.js';
+import { createElement } from '../../../global/render.js';
 
-function createTripMainInfoTemplate() {
-  return (/*html*/`
+const createTripMainInfoTemplate = () => (/*html*/`
     <section class="trip-main__trip-info  trip-info">
       <div class="trip-info__main">
         <h1 class="trip-info__title">Amsterdam &mdash; Chamonix &mdash; Geneva</h1>
@@ -12,8 +11,7 @@ function createTripMainInfoTemplate() {
         Total: &euro;&nbsp;<span class="trip-info__cost-value">1230</span>
       </p>
     </section>`
-  );
-}
+);
 
 export default class TripMainInfoView {
   getTemplate() {
